FV-7331 and FV-7313 fix score-fields and handle them as double. Fix evidence ID calculation

+++ fortscale/fortscale-services/src/main/java/fortscale/services/impl/EvidencesService.java
@@ -12,16 +12,16 @@ public interface EvidencesService {
 
 	/**
 	 * Create new object (transient) of evidence
-	 * @param entityType	The type of the entity
-	 * @param entityName	The name of the entity
-	 * @param date			The date of the evidence (single date for single event)
-	 * @param scoreFieldName	The field name of the anomaly
-	 * @param classifier		The relevant data source
-	 * @param score				The score
+	 * @param entityType    The type of the entity
+	 * @param entityName    The name of the entity
+	 * @param date            The date of the evidence (single date for single event)
+	 * @param scoreFieldName    The field name of the anomaly
+	 * @param classifier        The relevant data source
+	 * @param score                The score
 	 * @return	New evidence
 	 */
 	public Evidence createTransientEvidence(EntityType entityType, String entityName, Date date,
-			String scoreFieldName, String classifier, Integer score);
+			String scoreFieldName, String classifier, Double score);
 
 	/**
 	 * Create new evidence in Mongo
